backend/requirements.txt
fastapi
uvicorn
openai
gitpython
python-dotenv
